+++ baseframe/__init__.py
@@ -1,5 +1,5 @@
-from flask_babelhg import gettext as _
-from flask_babelhg import lazy_gettext as __
+from flask_babel import gettext as _
+from flask_babel import lazy_gettext as __
 
 from ._version import __version__, __version_info__
 from .assets import Version, assets